MiDAS example data

The microbial community composition of the activated sludge of 55 different Danish Wastewater Treatment Plants from 2006-2013 as determined by 16S rRNA amplicon sequencing using the Illumina platform.

data(MiDAS)

Format

An object of class "ampvis2". See amp_load.

Source

MiDAS: Field Guide to the Microbes of Activated Sludge and Anaerobic Digesters

References

McIlroy et al. (2015); MiDAS: the field guide to the microbes of activated sludge, Database, Volume 2015, bav062 (https://doi.org/10.1093/database/bav062)

Examples

data(MiDAS)
MiDAS

#> ampvis2 object with 5 elements.
#> Summary of OTU table:
#>      Samples         OTUs  Total#Reads    Min#Reads    Max#Reads Median#Reads 
#>          658        14969     20890850        10480        46264        31800 
#>    Avg#Reads 
#>     31749.01 
#> 
#> Assigned taxonomy:
#>       Kingdom        Phylum         Class         Order        Family 
#>   14969(100%) 14477(96.71%) 12737(85.09%) 11470(76.63%)  9841(65.74%) 
#>         Genus       Species 
#>   7380(49.3%)     28(0.19%) 
#> 
#> Metadata variables: 5 
#>  SampleID, Plant, Date, Year, Period