Generate a function table

Generate a function table that is sorted based on an input heatmap.

amp_function(heatmap)

Arguments

heatmap
(required) A heatmap generated by amp_heatmap.

functions
A data frame that associates functions with a genus names (default: "MiDAS").

adjust.bottom
Adjust the whitespace added in the bottom of the plot in mm (default: 10).

adjust.top
Adjust the whitespace added in the bottom of the plot in mm (default: 0).

genus.pos
Specify which part of the name that contains the genus name (default: 1).

point.size
Size of the plotted points (default: 6)

selection
A vector with the functions to be displayed. (default: c("MiDAS","FIL", "AOB", "NOB", "PAO", "GAO").

Value

A ggplot2 object.