Generate a heatmap of amplicon data

Generates a heatmap of amplicon data by using sample metadata to aggregate samples and taxonomy to aggregate OTUs.

amp_heatmap(data, group_by = "")

Arguments

data
(required) Data list as loaded with `amp_load()`.

group_by
A variable from the associated metadata to group the samples by.

facet_by
A variable from the associated metadata to facet the samples by.

normalise_by
A specific sample or variable in the metadata to normalise the counts by.

scale_by
A variable from the associated metadata to scale the abundance by.

tax_aggregate
The taxonomic level to aggregate the OTUs. (default: `"Phylum"`)

tax_add
Additional taxonomic level(s) to display, e.g. `"Phylum"`. (default: `"none"`)

tax_show
The number of taxa to show, or a vector of taxa names. (default: `10`)

tax_class
Converts a specific phylum to class level instead, e.g. `"p__Proteobacteria"`.

tax_empty
How to show OTUs without taxonomic information. One of the following: * `"remove"`: Remove OTUs without taxonomic information. * `"best"`: (default) Use the best classification possible. * `"OTU"`: Display the OTU name.

order_x_by
A taxonomy group or vector to order the x-axis by, or `"cluster"` for hierarchical clustering by `hclust()`.

order_y_by
A sample or vector to order the y-axis by, or `"cluster"` for hierarchical clustering by `hclust()`.

plot_values
(logical) Plot the values on the heatmap or not. (default: `TRUE`)

plot_breaks
A vector of breaks for the abundance legend, fx `c(1, 10, 20)`

plot_colorscale
The type of scale used for the coloring of abundances, either `"sqrt"` or `"log10"`. (default: `"log10"`)

plot_na
(logical) Whether to color missing values with the lowest color in the scale or not. (default: `TRUE`)

plot_values_size
The size of the plotted values. (default: `4`)

measure
Calculate and display either `"mean"`, `"max"` or `"median"` across the groups. (default: `"mean"`)

min_abundance
All values below this value are given the same color. (default: `0.1`)

max_abundance
All values above this value are given the same color.

sort_by
Sort the heatmap by a specific value of the `"group_by"` argument, e.g. `"Treatment A"`.

color_vector
Vector of colors for the colorscale, e.g. `c("white", "red")`.

round
Number of digits to show with the values. (default: `1`)

raw
(logical) Display raw input instead of converting to percentages. (default: `FALSE`)

detailed_output
(logical) Return additional details or not. If `TRUE`, it is recommended to save to an object and then access the additional data by `View(object$data)`. (default: `FALSE`)

Value

A ggplot2 object. If `detailed_output = TRUE` a list with a ggplot2 object and additional data.