Generates a ggplot2 style ordinate plot of amplicon data

A wrapper around the vegan package to make beautiful ggplot2 ordination plots to analyse and compare microbial community compositions. The input data must be a list loaded with amp_load() from the ampvis package, but any OTU table-like matrix (OTU's in rows, sampleID's in columns and abundances in the corresponding cells, aka a contingency table) alongside metadata for the same samples can be used. Simply choose an ordination type and a plot is returned.

amp_ordinate(data)

Arguments

data
(required) Data list loaded with amp_load() containing the elements: abund, metadata and tax.

filter_species
Remove low abundant OTU's across all samples below this threshold in percent. Recommended minimum: 0.1 pct (default: 0.1).

type
Ordination type; Principal Components Analysis(PCA), Redundancy Analysis(RDA), non-metric Multidimensional Scaling(NMDS), metric Multidimensional Scaling(MMDS, aka PCoA), Correspondence Analysis(CA), Canonical Correspondence Analysis(CCA) or Detrended Correspondence Analysis(DCA) (Default: PCA)

metric
Distance metric used for the distance-based ordination methods (nMDS/PCoA), any of the following: "manhattan", "euclidean", "canberra", "bray", "kulczynski", "jaccard", "gower", "jsd" (Jensen-Shannon Divergence), "altGower", "morisita", "horn", "mountford", "raup" , "binomial", "chao", "cao", "mahalanobis" or simply "none" or "sqrt". See details in ?vegdist, and for JSD http://enterotype.embl.de/enterotypes.html.

transform
Transform the abundance table with the decostand() function, fx "normalize", "chi.square", "hellinger" or "sqrt", see details in ?decostand. Using the hellinger transformation is always a good choice and is recommended for PCA/RDA/nMDS/PCoA to obtain a more ecologically meaningful result (learn about the double-zero problem).

constrain
Variable(s) in the metadata for constrained analyses (RDA and CCA). Multiple variables can be provided by a vector, fx c("Year", "Temperature"), but keep in mind the more variables the more the result will be similar to unconstrained analysis.

x_axis
Which axis from the ordination results to plot as the first axis. Have a look at the $screeplot with output = "detailed" to validate axes (default: 1).

y_axis
Which axis from the ordination results to plot as the second axis. Have a look at the $screeplot with output = "detailed" to validate axes (default: 2).

sample_color
Color sample points by a variable in the metadata.

sample_color_order
Order the colors in color by a vector.

sample_shape
Shape sample points by a variable in the metadata.

sample_colorframe
Frame the points with a polygon colored by the color argument (default: F).

sample_colorframe_label
Label by a variable in the metadata.

sample_label
Label sample points by a variable in the metadata.

sample_label_size
Sample labels text size (default: 4).

sample_label_segment_color
Sample labels repel-segment color (default: "black").

sample_trajectory
Make a trajectory between sample points by a variable in the metadata.

sample_trajectory_group
Make a trajectory between sample points by the trajectory argument, but within individual groups.

sample_plotly
Enable sample point howevering to display specific metadata or "all" (default: none).

species_plot
Plot species points (default: F).

species_nlabels
Number of most extreme species labels to plot.

species_label_taxonomy
Taxonomic level by which to label the species points (default: "Genus").

species_label_size
Size of the species text labels (default: 3).

species_label_color
Color of the species text labels (default: "grey10).

species_rescale
rescale species (default: F).

species_size
Size of the species points (default: 2).

species_shape
The shape of the points, fx 1 for hollow circles or 20 for dots (default: 20).

species_plotly
Enable species specific point howevering to display taxonomy (default: F).

envfit_factor
A vector of factor variables from the sample data used for envfit to the model

envfit_numeric
A vector of numerical variables from the sample data used for envfit to the model

envfit_signif_level
The significance treshold for displaying envfit parameters (default: 0.001).

envfit_textsize
Size of the envfit text on the plot (default: 3).

envfit_color
Color of the envfit text on the plot (default: "darkred").

envfit_numeric_arrows_scale
Scale the size of the numeric arrows (default: 1).

envfit_show
Show the results on the plot (default: T).

repel
Repel all labels to prevent cluttering of plots (default: T).

opacity
Opacity of all plotted points and sample_colorframe opacity, 0:invisible, 1:opaque (default: 0.8).

tax.empty
Option to add "best" classification or just the "OTU" name to each "OTU" (default: best).

output
"plot" or "detailed"; output as list with additional information(model, scores, inputmatrix etc) or just the plot (default: "plot).

...
Pass additional arguments to the vegan ordination functions, fx rda(...), cca(...), metaMDS(...), see vegan help.

Value

A ggplot2 object or a list with the ggplot2 object and associated dataframes.