Tidy up taxonomy

Used internally in other ampvis functions.

amp_rename(data, tax_class = NULL, tax_empty = "best",
  tax_level = "Genus")

Arguments

data
(required) Data list as loaded with amp_load.

tax_class
Converts a specific phylum to class level instead, e.g. "p__Proteobacteria".

tax_empty
How to show OTUs without taxonomic information. One of the following:

"remove": Remove OTUs without taxonomic information.

"best": (default) Use the best classification possible.

"OTU": Display the OTU name.

tax_level
The taxonomic level to remove OTUs with no assigned taxonomy, only used when tax_empty = "remove". (default: "Genus")

Value

A list with 3 dataframes (4 if reference sequences are provided).