Subset ampvis2 objects based on sample metadata

Subsets the data in ampvis2 objects based on metadata and returns the subsetted object.

amp_subset_samples(data, ...)

Arguments

data
(required) Data list as loaded with amp_load().

...
Logical expression indicating elements or rows to keep in the metadata. Missing values are taken as false. Directly passed to subset().

minreads
Minimum number of reads pr. sample. (default: 1)

normalise
(logical) Normalise the read abundances to the total amount of reads (percentages) BEFORE the subset. (default: FALSE)

Value

A list with 3 dataframes (4 if reference sequences are provided).

Details

The subset is performed on the metadata by subset() and the abundance- and taxonomy tables are then adjusted accordingly.

Examples

#Load example data
data("MiDAS")

#Show a short summary about the data by simply typing the name of the object in the console
MiDAS

#> ampvis2 object with 4 elements.
#> Summary of OTU table:
#>      Samples         OTUs  Total#Reads    Min#Reads    Max#Reads Median#Reads 
#>          573        14791     18398563        13306        46071        31758 
#>    Avg#Reads 
#>     32109.18 
#> 
#> Assigned taxonomy:
#>       Kingdom        Phylum         Class         Order        Family 
#>   14791(100%) 14283(96.57%) 12507(84.56%) 11184(75.61%)  9685(65.48%) 
#>         Genus       Species 
#>  7105(48.04%)     27(0.18%) 
#> 
#> Metadata variables:
#>  SampleID Plant Year Period Date DeltaDays LIB_ID Latitude Longitude RawData

#Keep only samples containing Aalborg West or East in the Plant column
MiDASsubset <- amp_subset_samples(MiDAS, Plant %in% c("Aalborg West", "Aalborg East"))

#> 522 samples and 6569 OTUs have been filtered 
#> Before: 573 samples and 14791 OTUs
#> After: 51 samples and 8222 OTUs

#Summary
MiDASsubset

#> ampvis2 object with 4 elements.
#> Summary of OTU table:
#>      Samples         OTUs  Total#Reads    Min#Reads    Max#Reads Median#Reads 
#>           51         8222      1498273        16751        38496        29230 
#>    Avg#Reads 
#>      29377.9 
#> 
#> Assigned taxonomy:
#>      Kingdom       Phylum        Class        Order       Family        Genus 
#>   8222(100%) 8037(97.75%) 7191(87.46%) 6597(80.24%) 5897(71.72%) 4532(55.12%) 
#>      Species 
#>    17(0.21%) 
#> 
#> Metadata variables:
#>  SampleID Plant Year Period Date DeltaDays LIB_ID Latitude Longitude RawData

#Keep only samples containing Aalborg West or East in the Plant column 
#and remove the sample "16SAMP-749". Remove any sample(s) with less than 20000 total reads
MiDASsubset2 <- amp_subset_samples(MiDAS,
    Plant %in% c("Aalborg West", "Aalborg East") & !SampleID %in% c("16SAMP-749"),
    minreads = 20000)

#> 525 samples and 6602 OTUs have been filtered 
#> Before: 573 samples and 14791 OTUs
#> After: 48 samples and 8189 OTUs

#Summary
MiDASsubset2

#> ampvis2 object with 4 elements.
#> Summary of OTU table:
#>      Samples         OTUs  Total#Reads    Min#Reads    Max#Reads Median#Reads 
#>           48         8189      1437013        20994        38496      29538.5 
#>    Avg#Reads 
#>     29937.77 
#> 
#> Assigned taxonomy:
#>      Kingdom       Phylum        Class        Order       Family        Genus 
#>   8189(100%) 8005(97.75%) 7163(87.47%) 6573(80.27%) 5875(71.74%) 4516(55.15%) 
#>      Species 
#>    17(0.21%) 
#> 
#> Metadata variables:
#>  SampleID Plant Year Period Date DeltaDays LIB_ID Latitude Longitude RawData