Subset ampvis2 objects based on taxonomy

Subsets the data in ampvis2 objects based on taxonomy and returns the subsetted object. See examples.

amp_subset_taxa(data, tax_vector)

Arguments

data
(required) Data list as loaded with amp_load.

tax_vector
(required) A vector with taxonomic groups, e.g. c("p__Chloroflexi","p__Actinobacteria").

normalise
(logical) Normalise the read abundances to the total amount of reads (percentages) BEFORE the subset. (default: FALSE)

Value

A list with 3 dataframes (4 if reference sequences are provided).

Details

The subset of taxa is done by providing a vector of taxa names, where all other taxa are then removed. The taxa names in the vector will be matched in all columns in the taxonomy table.

Normalising data for use in heatmaps

By default the raw read counts in the abundance matrix are normalised (transformed to percentages) by amp_heatmap automatically. This means that the relative abundances shown will be calculated based on the remaining taxa after the subset, not including the removed taxa, if any. To circumvent this, set normalise = TRUE when subsetting with the amp_subset_taxa and amp_subset_samples functions and then set raw = TRUE when using amp_heatmap, see the example below.

data("AalborgWWTPs")
subsettedData <- amp_subset_taxa(AalborgWWTPs,
                                 tax_vector = c("p__Chloroflexi", "p__Actinobacteria"),
                                 normalise = TRUE
                                 )
amp_heatmap(subsettedData,
            group_by = "Plant",
            tax_aggregate = "Phylum",
            tax_add = "Genus",
            raw = TRUE
            )

See also

amp_subset_samples, amp_heatmap

Examples

#Load example data
data("MiDAS")

#Show a short summary about the data by simply typing the name of the object in the console
MiDAS

#> ampvis2 object with 4 elements.
#> Summary of OTU table:
#>      Samples         OTUs  Total#Reads    Min#Reads    Max#Reads Median#Reads 
#>          573        14791     18398563        13306        46071        31758 
#>    Avg#Reads 
#>     32109.18 
#> 
#> Assigned taxonomy:
#>       Kingdom        Phylum         Class         Order        Family 
#>   14791(100%) 14283(96.57%) 12507(84.56%) 11184(75.61%)  9685(65.48%) 
#>         Genus       Species 
#>  7105(48.04%)     27(0.18%) 
#> 
#> Metadata variables: 10 
#>  SampleID, Plant, Year, Period, Date, DeltaDays, LIB_ID, Latitude, Longitude, RawData

#Remove all taxa except the phyla Chloroflexi and Actinobacteria
MiDASsubset <- amp_subset_taxa(MiDAS, tax_vector = c("p__Chloroflexi", "p__Actinobacteria"))

#Summary
MiDASsubset

#> ampvis2 object with 4 elements.
#> Summary of OTU table:
#>      Samples         OTUs  Total#Reads    Min#Reads    Max#Reads Median#Reads 
#>          573         2752      7041712         2898        28818        11706 
#>    Avg#Reads 
#>      12289.2 
#> 
#> Assigned taxonomy:
#>      Kingdom       Phylum        Class        Order       Family        Genus 
#>   2752(100%)   2752(100%) 2666(96.88%) 2490(90.48%) 2102(76.38%) 1644(59.74%) 
#>      Species 
#>        0(0%) 
#> 
#> Metadata variables: 10 
#>  SampleID, Plant, Year, Period, Date, DeltaDays, LIB_ID, Latitude, Longitude, RawData