Timeseries

This function allows you plot relative abundance reads over time in a timeseries plot.

amp_timeseries(data)

Arguments

data
(required) Data list as loaded with `amp_load()`.

time
(required) The name of the column containing the time variables, e.g. "Date".

group
A variable from the associated sample data to group/split samples by (default: "Samples").

split
Split the plot into subplot of each taxa (default: F).

tax.show
The number of taxa to show or a vector of taxa names (default: 5).

tax.aggregate
The taxonomic level that the data should be aggregated to (default: "OTU")

tax.add
Additional taxonomic levels to display for each entry, e.g. "Phylum" (default: "none").

tax.class
Converts a specific phyla to class level instead, e.g. "p__Proteobacteria" (default: "none) .

tax.empty
Either "remove" OTUs without taxonomic information, add "best" classification or add the "OTU" name (default: "best").

raw
Display raw input instead of converting to percentages (default: F)

Value

A ggplot2 object or a list with the ggplot2 object and associated dataframes.