Transform ampvis2 objects into a long-format data frame

Transforms any ampvis2 object into a long data frame (data.table) to facilitate custom data analysis. Only the elements OTU counts (data$abund), taxonomy (data$tax), and metadata (data$metadata) are used, not phylogenetic tree or DNA sequences.

amp_export_long(
  data,
  metadata_vars = colnames(data$metadata),
  tax_levels = colnames(data$tax)
)

Arguments

data
(required) Data list as loaded with amp_load.

metadata_vars
A character vector of sample metadata variables to include from data$metadata. The first column (sample IDs) will always be included. Default includes all.

tax_levels
A character vector of taxonomic levels to include from data$tax. The OTU column will always be included. Default includes all.

Value

A data.table in long format.

Author

Kasper Skytte Andersen ksa@bio.aau.dk

Thomas Yssing Michaelsen tym@bio.aau.dk

Examples

# load minimal example data
d <- amp_load(example_otutable, example_metadata)

# transform d into a long-format data frame
d_long <- amp_export_long(d, metadata_vars = NULL, tax_levels = c("OTU", "Genus"))

# print the data frame (data.table)
d_long
#>        SampleID    OTU count                       Genus
#>  1: 16SAMP_3893  OTU_1    23    g__Candidatus Amarilinum
#>  2: 16SAMP_3893 OTU_10   373                     g__MK04
#>  3: 16SAMP_3893  OTU_2   675             g__Tetrasphaera
#>  4: 16SAMP_3893  OTU_3   780    g__Candidatus Microthrix
#>  5: 16SAMP_3893  OTU_4   272            g__Dechloromonas
#>  6: 16SAMP_3893  OTU_5   560 g__Candidatus Villogracilis
#>  7: 16SAMP_3893  OTU_6   906             g__Trichococcus
#>  8: 16SAMP_3893  OTU_7   297    g__Candidatus Microthrix
#>  9: 16SAMP_3893  OTU_8    28               g__Nitrospira
#> 10: 16SAMP_3893  OTU_9     0                g__QEDR3BF09
#> 11: 16SAMP_3913  OTU_1    15    g__Candidatus Amarilinum
#> 12: 16SAMP_3913 OTU_10   256                     g__MK04
#> 13: 16SAMP_3913  OTU_2   565             g__Tetrasphaera
#> 14: 16SAMP_3913  OTU_3   733    g__Candidatus Microthrix
#> 15: 16SAMP_3913  OTU_4   233            g__Dechloromonas
#> 16: 16SAMP_3913  OTU_5   339 g__Candidatus Villogracilis
#> 17: 16SAMP_3913  OTU_6   766             g__Trichococcus
#> 18: 16SAMP_3913  OTU_7   218    g__Candidatus Microthrix
#> 19: 16SAMP_3913  OTU_8     8               g__Nitrospira
#> 20: 16SAMP_3913  OTU_9     0                g__QEDR3BF09
#> 21: 16SAMP_3941  OTU_1   273    g__Candidatus Amarilinum
#> 22: 16SAMP_3941 OTU_10    19                     g__MK04
#> 23: 16SAMP_3941  OTU_2   331             g__Tetrasphaera
#> 24: 16SAMP_3941  OTU_3   405    g__Candidatus Microthrix
#> 25: 16SAMP_3941  OTU_4  1434            g__Dechloromonas
#> 26: 16SAMP_3941  OTU_5   509 g__Candidatus Villogracilis
#> 27: 16SAMP_3941  OTU_6   133             g__Trichococcus
#> 28: 16SAMP_3941  OTU_7   418    g__Candidatus Microthrix
#> 29: 16SAMP_3941  OTU_8   155               g__Nitrospira
#> 30: 16SAMP_3941  OTU_9     9                g__QEDR3BF09
#> 31: 16SAMP_3946  OTU_1    51    g__Candidatus Amarilinum
#> 32: 16SAMP_3946 OTU_10   415                     g__MK04
#> 33: 16SAMP_3946  OTU_2   411             g__Tetrasphaera
#> 34: 16SAMP_3946  OTU_3   199    g__Candidatus Microthrix
#> 35: 16SAMP_3946  OTU_4   256            g__Dechloromonas
#> 36: 16SAMP_3946  OTU_5   598 g__Candidatus Villogracilis
#> 37: 16SAMP_3946  OTU_6   390             g__Trichococcus
#> 38: 16SAMP_3946  OTU_7   130    g__Candidatus Microthrix
#> 39: 16SAMP_3946  OTU_8    72               g__Nitrospira
#> 40: 16SAMP_3946  OTU_9     0                g__QEDR3BF09
#> 41: 16SAMP_3953  OTU_1   127    g__Candidatus Amarilinum
#> 42: 16SAMP_3953 OTU_10    43                     g__MK04
#> 43: 16SAMP_3953  OTU_2   430             g__Tetrasphaera
#> 44: 16SAMP_3953  OTU_3  1346    g__Candidatus Microthrix
#> 45: 16SAMP_3953  OTU_4   736            g__Dechloromonas
#> 46: 16SAMP_3953  OTU_5   223 g__Candidatus Villogracilis
#> 47: 16SAMP_3953  OTU_6   232             g__Trichococcus
#> 48: 16SAMP_3953  OTU_7  1354    g__Candidatus Microthrix
#> 49: 16SAMP_3953  OTU_8   156               g__Nitrospira
#> 50: 16SAMP_3953  OTU_9    19                g__QEDR3BF09
#> 51: 16SAMP_4591  OTU_1   190    g__Candidatus Amarilinum
#> 52: 16SAMP_4591 OTU_10   102                     g__MK04
#> 53: 16SAMP_4591  OTU_2   780             g__Tetrasphaera
#> 54: 16SAMP_4591  OTU_3  1114    g__Candidatus Microthrix
#> 55: 16SAMP_4591  OTU_4  1338            g__Dechloromonas
#> 56: 16SAMP_4591  OTU_5   145 g__Candidatus Villogracilis
#> 57: 16SAMP_4591  OTU_6  1458             g__Trichococcus
#> 58: 16SAMP_4591  OTU_7   198    g__Candidatus Microthrix
#> 59: 16SAMP_4591  OTU_8   101               g__Nitrospira
#> 60: 16SAMP_4591  OTU_9    25                g__QEDR3BF09
#> 61: 16SAMP_4597  OTU_1   220    g__Candidatus Amarilinum
#> 62: 16SAMP_4597 OTU_10    73                     g__MK04
#> 63: 16SAMP_4597  OTU_2   699             g__Tetrasphaera
#> 64: 16SAMP_4597  OTU_3  1630    g__Candidatus Microthrix
#> 65: 16SAMP_4597  OTU_4  1224            g__Dechloromonas
#> 66: 16SAMP_4597  OTU_5   212 g__Candidatus Villogracilis
#> 67: 16SAMP_4597  OTU_6   560             g__Trichococcus
#> 68: 16SAMP_4597  OTU_7   283    g__Candidatus Microthrix
#> 69: 16SAMP_4597  OTU_8   151               g__Nitrospira
#> 70: 16SAMP_4597  OTU_9    58                g__QEDR3BF09
#> 71: 16SAMP_4603  OTU_1    83    g__Candidatus Amarilinum
#> 72: 16SAMP_4603 OTU_10   138                     g__MK04
#> 73: 16SAMP_4603  OTU_2   820             g__Tetrasphaera
#> 74: 16SAMP_4603  OTU_3   112    g__Candidatus Microthrix
#> 75: 16SAMP_4603  OTU_4   564            g__Dechloromonas
#> 76: 16SAMP_4603  OTU_5  1619 g__Candidatus Villogracilis
#> 77: 16SAMP_4603  OTU_6   287             g__Trichococcus
#> 78: 16SAMP_4603  OTU_7   116    g__Candidatus Microthrix
#> 79: 16SAMP_4603  OTU_8    25               g__Nitrospira
#> 80: 16SAMP_4603  OTU_9     0                g__QEDR3BF09
#>        SampleID    OTU count                       Genus