Generate a heatmap from amplicon data

Generate a heatmap in ggplot2 format from amplicon data in ampvis format. Use sample metadata to aggregate samples and taxonomy to aggregate OTUs.

amp_heatmap(data)

Arguments

data
(required) A ampvis object.

group
A variable from the associated sample data to group samples by.

facet
A variable to facet the plot by.

normalise
A specific sample or group to normalise the counts to, or "relative".

scale
A variable from the associated sample data to scale the abundance by.

tax.aggregate
The taxonomic level that the data should be aggregated to (default: "Phylum").

tax.add
Additional taxonomic levels to display for each entry e.g. "Phylum" (default: "none").

tax.show
The number of taxa to show or a vector of taxa names (default: 10).

tax.class
Converts a specific phyla to class level instead (e.g. "p__Proteobacteria").

tax.empty
Either "remove" OTUs without taxonomic information, add "best" classification or add the "OTU" name (default: "best").

order.x
A taxonomy group or vector to order the x-axis by, alternatively "cluster".

order.y
A sample or vector to order the y-axis by, alternatively "cluster".

plot.numbers
Plot the values on the heatmap (default: T)

plot.breaks
A vector of breaks for the abundance legend.

plot.colorscale
Either "sqrt" or "log10" (default: "log10")

plot.na
Whether to color missing values with the lowest color in the scale (default: T).

output
To output a plot or the complete data inclusive dataframes (default: "plot")

plot.text.size
The size of the plotted text (default: 4).

calc
Calculate and display "mean", "max" or "median" across the groups (default: "mean").

min.abundance
All values below are given the same color (default: 0.1).

max.abundance
All values above are given the same color.

sort.by
Sort the heatmap by a specific value of the "group", e.g. "Treatment A".

color.vector
Vector with colors for colorscale e.g. c("red","white") (default: NULL)

round
Number of digits to plot (default: 1)

raw
Display raw input instead of converting to percentages (default: F)

Value

A ggplot2 object or a list with the ggplot2 object and associated dataframes.