Load data for ampvis functions

This function reads an OTU-table and corresponding sample metadata, and returns a list for use in all ampvis functions. It is therefore required to load data with amp_load() before any other ampvis functions can be used.

amp_load(otutable = dataframe, metadata = dataframe)

Arguments

otutable
(required) An OTU-table (data frame), where the last 7 rows is the taxonomy.

metadata
(required) A metadata (dataframe) with information about the samples.

refseq
Reference sequences for all OTUs as loaded with readDNAStringSet() from the Biostrings bioconductor package.

Value

A list with 3 dataframes (4 if reference sequences are provided).

Details

The amp_load() function validates and corrects the provided data frames in different ways to make it suitable for the rest of the ampvis functions. It is important that the provided data frames match the requirements as described in the following sections to work properly.

The OTU-table

The OTU-table contains information about the OTUs, their assigned taxonomy and their read counts in each sample. The provided OTU-table must be a data frame with the following requirements:

The rows are OTU IDs and the columns are samples.

The last 7 columns are the corresponding taxonomy assigned to the OTUs, named "Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species".

The rownames of the data frame are the OTU IDs. Thus, the first column should be a sample, NOT the OTU IDs.

The column names of the data frame are the sample IDs, exactly matching those in the metadata, (and the last 7 columns named Kingdom -> Species, of course). Thus, the first row should contain the read counts of one of the OTUs in each sample, NOT the sample IDs and taxonomy.

Generally avoid special characters and spaces in row- and column names.

The metadata

The metadata contains additional information about the samples, for example where each sample was taken, date, pH, treatment etc, which is used to compare and group the samples during analysis. The amount of information in the metadata is unlimited, it can contain any number of columns (variables), however there are a few requirements:

The sample IDs must be in the first column. These sample IDs must match exactly to those in the OTU-table.

Column classes matter, categorical variables should be loaded either as.character() or as.factor(), and continuous variables as.numeric(). See below.

Generally avoid special characters and spaces in row- and column names.

If for example a column is named "Year" and the entries are simply entered as numbers (2011, 2012, 2013 etc), then R will automatically consider these as numerical values (as.numeric()) and therefore the column as a continuous variable, while it is a categorical variable and should be loaded as.factor() or as.character() instead. This has consequences for the analysis as R treats them differently. Therefore either use the colClasses = argument when loading a csv file or col_types = when loading an excel file, or manually adjust the column classes afterwards with fx metadata$Year <- as.character(metadata$Year). amp_load() will automatically use the sample IDs in the first column as rownames, but it is important to also have an actual column with sample IDs, so it is possible to fx group by that column during analysis.

Examples

#Be sure to use the correct function to load your .csv files, see ?read.table()

# NOT RUN {
myotutable <- read.delim("data/otutable.txt",
                        header = TRUE,
                        fill = TRUE,
                        stringsAsFactors = FALSE,
                        check.names = FALSE,
                        row.names = 1) %>% as.data.frame()
mymetadata <- read_excel("data/metadata.xlsx",
                         col_names = TRUE) %>% as.data.frame()
d <- amp_load(otutable = myotutable,
              metadata = mymetadata
              #refseq is optional:
              #, refseq = readDNAStringSet("data/otus.fa", format = "fasta") 
              )
d
# }