Timeseries

Generates a timeseries plot showing relative read abundances over time.

amp_timeseries(data, time_variable = "")

Arguments

data
(required) Data list as loaded with amp_load().

time_variable
(required) The name of the column in the metadata containing the time variables, e.g. "Date". Must be directly compatible with as.Date() and preferably of the form "yyyy-mm-dd" or "%Y-%m-%d".

group_by
Group the samples by a variable in the metadata.

split
Split the plot into subplots of each taxa. (default: FALSE)

tax_show
The number of taxa to show, or a vector of taxa names. (default: 6)

tax_aggregate
The taxonomic level to aggregate the OTUs. (default: "OTU")

tax_add
Additional taxonomic level(s) to display, e.g. "Phylum". (default: "none")

tax_class
Converts a specific phylum to class level instead, e.g. "p__Proteobacteria".

tax_empty
How to show OTUs without taxonomic information. One of the following:

"remove": Remove OTUs without taxonomic information.

"best": (default) Use the best classification possible.

"OTU": Display the OTU name.

raw
(logical) Display raw input instead of converting to percentages. (default: FALSE)

...
Additional arguments passed to as.Date() to make the time_variable compatible with the timeseries plot, fx the format or tz arguments, see ?as.Date.

Value

A ggplot2 object.

Examples

data("AalborgWWTPs")
amp_timeseries(AalborgWWTPs, time_variable = "Date")

#> Error in `:=`("sum", sum(Abundance)): Check that is.data.table(DT) == TRUE. Otherwise, := and `:=`(...) are defined for use in j, once only and in particular ways. See help(":=").

amp_timeseries(AalborgWWTPs, time_variable = "Date", split = TRUE, tax_aggregate = "Genus")

#> Error in `:=`("sum", sum(Abundance)): Check that is.data.table(DT) == TRUE. Otherwise, := and `:=`(...) are defined for use in j, once only and in particular ways. See help(":=").